Supress warning
`DeprecationWarning: The default dtype for empty Series will be 'object' instead of 'float64' in a future version. Specify a dtype explicitly to silence this warning.`

+++ pandas/tests/generic/test_to_xarray.py
@@ -99,7 +99,7 @@ class TestSeriesToXArray:
 
         from xarray import DataArray
 
-        s = Series(range(len(indices)), index=indices)
+        s = Series(range(len(indices)), index=indices, dtype="float64")
         s.index.name = "foo"
         result = s.to_xarray()
         repr(result)
